Alternative upEx codegen
The old exception handling codegen had to create
an exception and get the stack trace to determine
the stack depth.  This is a problem with deep
stacks, as the default stack depth is 1000
elements and thus we had to specify some arbitrary
limit like -XX:MaxJavaStackTraceDepth=100000 to
support deep stacks.

With this patch, each method that has exception
handlers for Pausable methods loads the current 
stack depth into a local value in the preamble;
and thus upEx does not need to get a stack trace
to determine the stack depth.  This also makes
exceptions faster, since the stack traces don't 
need reifying.